moved common functions to common file
