+++ airstack-modules/modules/airstack/common/index.ts
@@ -1,7 +1,13 @@
 import {
   BigInt,
-  TypedMap
+  TypedMap,
+  dataSource,
 } from "@graphprotocol/graph-ts";
+import {
+  AirBlock,
+  AirEntityCounter,
+  AirMeta,
+} from "../../../generated/schema";
 
 export const AIR_META_ID = "AIR_META";
 
@@ -42,4 +48,80 @@ export function processNetwork(network: string): string {
   const value = AIR_NETWORK_MAP.get(network);
   const result: string = value !== null ? value : "unknown";
   return result;
+}
+
+//air entity funcitons
+
+/**
+ * @dev this function updates air entity counter for a given entity id
+ * @param id entity id for entity to be updated
+ * @param block air block object
+ * @returns updated entity count
+ */
+export function updateAirEntityCounter(
+  id: string,
+  block: AirBlock,
+): BigInt {
+  let entity = AirEntityCounter.load(id);
+  if (entity == null) {
+    entity = new AirEntityCounter(id);
+    entity.count = BIGINT_ONE;
+    entity.createdAt = block.id;
+    entity.lastUpdatedAt = block.id;
+    createAirMeta(SUBGRAPH_SLUG, SUBGRAPH_NAME);
+  } else {
+    entity.count = entity.count.plus(BIGINT_ONE);
+    entity.lastUpdatedAt = block.id;
+  }
+  entity.save();
+  return entity.count as BigInt;
+}
+
+/**
+ * @dev this function creates air meta entity
+ * @param slug subgraph slug
+ * @param name subgraph name
+ */
+export function createAirMeta(
+  slug: string,
+  name: string
+  // should ideally have version also being passed from here
+): void {
+  let meta = AirMeta.load(AIR_META_ID);
+  if (meta == null) {
+    meta = new AirMeta(AIR_META_ID);
+    meta.network = processNetwork(dataSource.network());
+    meta.schemaVersion = SUBGRAPH_SCHEMA_VERSION;
+    meta.version = SUBGRAPH_VERSION;
+    meta.slug = slug;
+    meta.name = name;
+    meta.save();
+  }
+}
+
+/**
+ * @dev this function gets or creates a new air block entity
+ * @param chainId chain id
+ * @param blockHeight block number
+ * @param blockHash block hash
+ * @param blockTimestamp block timestamp
+ * @returns AirBlock entity
+ */
+export function getOrCreateAirBlock(
+  chainId: string,
+  blockHeight: BigInt,
+  blockHash: string,
+  blockTimestamp: BigInt
+): AirBlock {
+  let id = chainId.concat("-").concat(blockHeight.toString());
+
+  let block = AirBlock.load(id);
+  if (block == null) {
+    block = new AirBlock(id);
+    block.hash = blockHash;
+    block.number = blockHeight;
+    block.timestamp = blockTimestamp
+    block.save()
+  }
+  return block as AirBlock;
 }
