+++ ens/modules/airstack/domain-name/domain-name.ts
@@ -1,6 +1,5 @@
 import {
   BigInt,
-  dataSource,
   Bytes,
   crypto,
   log,
@@ -10,10 +9,8 @@ import {
 
 import {
   AirAccount,
-  AirEntityCounter,
   AirToken,
   AirBlock,
-  AirMeta,
   AirDomain,
   AirDomainOwnerChangedTransaction,
   AirDomainTransferTransaction,
@@ -24,8 +21,8 @@ import {
   AirNameRenewedTransaction,
   AirAddrChanged,
 } from "../../../generated/schema";
-import { uint256ToByteArray, AIR_META_ID, AIR_DOMAIN_OWNER_CHANGED_ENTITY_COUNTER_ID, AIR_ADDR_CHANGED_TRANSACTION_COUNTER_ID, AIR_NAME_RENEWED_TRANSACTION_COUNTER_ID, AIR_NAME_REGISTERED_TRANSACTION_COUNTER_ID, AIR_DOMAIN_NEW_TTL_TRANSACTION_COUNTER_ID, AIR_DOMAIN_NEW_RESOLVER_ENTITY_COUNTER_ID, AIR_DOMAIN_TRANSFER_ENTITY_COUNTER_ID, ROOT_NODE, ZERO_ADDRESS } from "./utils";
-import { BIGINT_ONE, SUBGRAPH_SCHEMA_VERSION, SUBGRAPH_VERSION, SUBGRAPH_NAME, SUBGRAPH_SLUG, processNetwork, BIG_INT_ZERO, EMPTY_STRING } from "../common";
+import { uint256ToByteArray, AIR_DOMAIN_OWNER_CHANGED_ENTITY_COUNTER_ID, AIR_ADDR_CHANGED_TRANSACTION_COUNTER_ID, AIR_NAME_RENEWED_TRANSACTION_COUNTER_ID, AIR_NAME_REGISTERED_TRANSACTION_COUNTER_ID, AIR_DOMAIN_NEW_TTL_TRANSACTION_COUNTER_ID, AIR_DOMAIN_NEW_RESOLVER_ENTITY_COUNTER_ID, AIR_DOMAIN_TRANSFER_ENTITY_COUNTER_ID, ROOT_NODE, ZERO_ADDRESS } from "./utils";
+import { BIGINT_ONE, BIG_INT_ZERO, EMPTY_STRING, updateAirEntityCounter, getOrCreateAirBlock } from "../common";
 
 export namespace domain {
   /**
@@ -946,80 +943,6 @@ export namespace domain {
       return null
     }
     return domain.id
-  }
-
-  /**
-   * @dev this function updates air entity counter for a given entity id
-   * @param id entity id for entity to be updated
-   * @param block air block object
-   * @returns updated entity count
-   */
-  function updateAirEntityCounter(
-    id: string,
-    block: AirBlock,
-  ): BigInt {
-    let entity = AirEntityCounter.load(id);
-    if (entity == null) {
-      entity = new AirEntityCounter(id);
-      entity.count = BIGINT_ONE;
-      entity.createdAt = block.id;
-      entity.lastUpdatedAt = block.id;
-      createAirMeta(SUBGRAPH_SLUG, SUBGRAPH_NAME);
-    } else {
-      entity.count = entity.count.plus(BIGINT_ONE);
-      entity.lastUpdatedAt = block.id;
-    }
-    entity.save();
-    return entity.count as BigInt;
-  }
-
-  /**
-   * @dev this function creates air meta entity
-   * @param slug subgraph slug
-   * @param name subgraph name
-   */
-  function createAirMeta(
-    slug: string,
-    name: string
-    // should ideally have version also being passed from here
-  ): void {
-    let meta = AirMeta.load(AIR_META_ID);
-    if (meta == null) {
-      meta = new AirMeta(AIR_META_ID);
-      meta.network = processNetwork(dataSource.network());
-      meta.schemaVersion = SUBGRAPH_SCHEMA_VERSION;
-      meta.version = SUBGRAPH_VERSION;
-      meta.slug = slug;
-      meta.name = name;
-      meta.save();
-    }
-  }
-
-  /**
-   * @dev this function gets or creates a new air block entity
-   * @param chainId chain id
-   * @param blockHeight block number
-   * @param blockHash block hash
-   * @param blockTimestamp block timestamp
-   * @returns AirBlock entity
-   */
-  function getOrCreateAirBlock(
-    chainId: string,
-    blockHeight: BigInt,
-    blockHash: string,
-    blockTimestamp: BigInt
-  ): AirBlock {
-    let id = chainId.concat("-").concat(blockHeight.toString());
-
-    let block = AirBlock.load(id);
-    if (block == null) {
-      block = new AirBlock(id);
-      block.hash = blockHash;
-      block.number = blockHeight;
-      block.timestamp = blockTimestamp
-      block.save()
-    }
-    return block as AirBlock;
   }
 
   /**
